Moved column visibility reducer to its own file; fixed typo
